// src/generated/axios/api.ts
export interface DoctorDTO {
    /**
     * 
     * @type {number}
     * @memberof DoctorDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'surname'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'phoneNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'address'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'avatar'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorDTO
     */
    'profession'?: string;
    /**
     * 
     * @type {Array<PatientDTO>}
     * @memberof DoctorDTO
     */
    'latestPatients'?: Array<PatientDTO>;
}
/**
 * 
 * @export
 * @interface DoctorFilterDTO
 */
export interface DoctorFilterDTO {
    /**
     * 
     * @type {string}
     * @memberof DoctorFilterDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorFilterDTO
     */
    'surname'?: string;
    /**
     * 
     * @type {string}
     * @memberof DoctorFilterDTO
     */
    'profession'?: string;
}
/**
 * 
 * @export
 * @interface PatientDTO
 */
export interface PatientDTO {
    /**
     * 
     * @type {number}
     * @memberof PatientDTO
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientDTO
     */
    'opd'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientDTO
     */
    'idp'?: number;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'surname'?: string;
    /**
     * 
     * @type {number}
     * @memberof PatientDTO
     */
    'phoneNumber'?: number;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'address'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'avatar'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'bloodGroup'?: PatientDTOBloodGroupEnum;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'notes'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof PatientDTO
     */
    'chronicPatient'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PatientDTO
     */
    'lastAdmission'?: string;
    /**
     * 
     * @type {number}
     * @memberof PatientDTO
     */
    'lastDoctorVisitedId'?: number;
    /**
     * 
     * @type {Array<PatientRecordDTO>}
     * @memberof PatientDTO
     */
    'patientRecords'?: Array<PatientRecordDTO>;
    /**
     * 
     * @type {Array<number>}
     * @memberof PatientDTO
     */
    'doctorIds'?: Array<number>;
}

export const PatientDTOBloodGroupEnum = {
    ZeroPlus: 'ZERO_PLUS',
    ZeroMinus: 'ZERO_MINUS',
    APlus: 'A_PLUS',
    AMinus: 'A_MINUS',
    BPlus: 'B_PLUS',
    BMinus: 'B_MINUS',
    AbPlus: 'AB_PLUS',
    AbMinus: 'AB_MINUS'
} as const;

export type PatientDTOBloodGroupEnum = typeof PatientDTOBloodGroupEnum[keyof typeof PatientDTOBloodGroupEnum];

/**
 * 
 * @export
 * @interface PatientFilterDTO
 */
export interface PatientFilterDTO {
    /**
     * 
     * @type {string}
     * @memberof PatientFilterDTO
     */
    'text'?: string;
    /**
     * 
     * @type {number}
     * @memberof PatientFilterDTO
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientFilterDTO
     */
    'opd'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientFilterDTO
     */
    'idp'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientFilterDTO
     */
    'doctorId'?: number;
}
/**
 * 
 * @export
 * @interface PatientRecordDTO
 */
export interface PatientRecordDTO {
    /**
     * 
     * @type {number}
     * @memberof PatientRecordDTO
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof PatientRecordDTO
     */
    'patientId'?: number;
    /**
     * 
     * @type {DoctorDTO}
     * @memberof PatientRecordDTO
     */
    'doctor'?: DoctorDTO;
    /**
     * 
     * @type {string}
     * @memberof PatientRecordDTO
     */
    'date'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientRecordDTO
     */
    'typeVisit'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientRecordDTO
     */
    'reasonVisit'?: string;
    /**
     * 
     * @type {string}
     * @memberof PatientRecordDTO
     */
    'treatmentMade'?: string;
}
// ...
export class DoctorApiApi extends BaseAPI {
    /**
     * 
     * @param {DoctorDTO} doctorDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DoctorApiApi
     */
    public createDoctor(doctorDTO: DoctorDTO, options?: RawAxiosRequestConfig) {
        return DoctorApiApiFp(this.configuration).createDoctor(doctorDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DoctorApiApi
     */
    public deleteDoctor(id: number, options?: RawAxiosRequestConfig) {
        return DoctorApiApiFp(this.configuration).deleteDoctor(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DoctorApiApi
     */
    public getDoctor(id: number, options?: RawAxiosRequestConfig) {
        return DoctorApiApiFp(this.configuration).getDoctor(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} page 
     * @param {number} size 
     * @param {string} sort 
     * @param {DoctorFilterDTO} doctorFilterDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DoctorApiApi
     */
    public getListDoctor(page: number, size: number, sort: string, doctorFilterDTO: DoctorFilterDTO, options?: RawAxiosRequestConfig) {
        return DoctorApiApiFp(this.configuration).getListDoctor(page, size, sort, doctorFilterDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {DoctorDTO} doctorDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DoctorApiApi
     */
    public updateDoctor(id: number, doctorDTO: DoctorDTO, options?: RawAxiosRequestConfig) {
        return DoctorApiApiFp(this.configuration).updateDoctor(id, doctorDTO, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class PatientApiApi extends BaseAPI {
    /**
     * 
     * @param {PatientDTO} patientDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientApiApi
     */
    public createPatient(patientDTO: PatientDTO, options?: RawAxiosRequestConfig) {
        return PatientApiApiFp(this.configuration).createPatient(patientDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientApiApi
     */
    public deletePatient(id: number, options?: RawAxiosRequestConfig) {
        return PatientApiApiFp(this.configuration).deletePatient(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} page 
     * @param {number} size 
     * @param {string} sort 
     * @param {PatientFilterDTO} patientFilterDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientApiApi
     */
    public getListPatient(page: number, size: number, sort: string, patientFilterDTO: PatientFilterDTO, options?: RawAxiosRequestConfig) {
        return PatientApiApiFp(this.configuration).getListPatient(page, size, sort, patientFilterDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientApiApi
     */
    public getPatient(id: number, options?: RawAxiosRequestConfig) {
        return PatientApiApiFp(this.configuration).getPatient(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {PatientDTO} patientDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientApiApi
     */
    public updatePatient(id: number, patientDTO: PatientDTO, options?: RawAxiosRequestConfig) {
        return PatientApiApiFp(this.configuration).updatePatient(id, patientDTO, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class PatientRecordApiApi extends BaseAPI {
    /**
     * 
     * @param {PatientRecordDTO} patientRecordDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientRecordApiApi
     */
    public createPatientRecord(patientRecordDTO: PatientRecordDTO, options?: RawAxiosRequestConfig) {
        return PatientRecordApiApiFp(this.configuration).createPatientRecord(patientRecordDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientRecordApiApi
     */
    public deletePatientRecord(id: number, options?: RawAxiosRequestConfig) {
        return PatientRecordApiApiFp(this.configuration).deletePatientRecord(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientRecordApiApi
     */
    public getPatientRecord(id: number, options?: RawAxiosRequestConfig) {
        return PatientRecordApiApiFp(this.configuration).getPatientRecord(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {PatientRecordDTO} patientRecordDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PatientRecordApiApi
     */
    public updatePatientRecord(id: number, patientRecordDTO: PatientRecordDTO, options?: RawAxiosRequestConfig) {
        return PatientRecordApiApiFp(this.configuration).updatePatientRecord(id, patientRecordDTO, options).then((request) => request(this.axios, this.basePath));
    }
}